basic PID is running

+++ ESP32C6_Controller/spider_main/main/encoder.h
@@ -5,9 +5,9 @@
 void encoder_init();
 void reset_encoder(int num);
 
-void get_counts(uint8_t *enc_pos);
+void get_counts(int *enc_pos);
 
 int get_count(int num);
-uint8_t get_angle(int num);
+int get_angle(int num);
 
 #endif
